diario okay

+++ bulletin/services/metabase.py
@@ -7,7 +7,7 @@ import urllib3
 import cgi
 from clint.textui import progress
 from os import makedirs, listdir
-from os.path import join, dirname, isdir
+from os.path import join, dirname, isdir, isfile
 from bulletin import root, default_input, default_output
 from retry import retry
 import re
@@ -16,15 +16,15 @@ from pathlib import Path
 
 from bulletin.utils.normalize import trim_overspace
 from bulletin.utils.timer import Timer
-from bulletin.notifica import Notifica
 
 urllib3.disable_warnings()
 
 class Metabase:
+
     def __init__(self, limit:int=2**18):
         self.limit = limit
 
-        self.cookie = self.__request_cookie()
+        self.cookie = self.request_cookie()
         self.output = join(default_input,'queries')
         self.sql_dir = join(root,'resources','sql')
         self.tables = join(root,'resources','tables')
@@ -40,16 +40,66 @@ class Metabase:
         
         self.saved_queries = [ Path(path).stem for path in glob.glob(join(self.sql_dir,"*.sql"))]
         
-        print("\nsaved_queries:",self.saved_queries)
+        print("\nsaved_queries:",self.saved_queries,'\n')
+    
+    
+    def request_cookie(self):
+        my_cookie = ''
+        if not isfile(join(default_input,'cookie_file')):
+            print('login_session')
+            raise Exception('not_implemented')
+            
+        with open(join(default_input,'cookie_file'),'r') as cookie_file:
+            my_cookie = cookie_file.read()
+        
+        print(f"Cookie:'{my_cookie[:5]}{''.join(['-' if x=='-' else '*' for x in my_cookie])}{my_cookie[-5:]}'")
+        return my_cookie#input('enter with cookie')                                                                 
+                                                             
     
     @Timer('Download tables')
     def download_tables(self, update=False):
         
-        notificacao = pd.read_csv(self.download("select table_name, column_name, udt_name from information_schema.columns where table_name = 'notificacao' order by table_name", join(self.tables,'notificacao.csv')))
+        notificacao_sql = '''
+            select 
+                table_name, column_name, udt_name 
+            from 
+                information_schema.columns 
+            where 
+                table_name = 'notificacao' 
+            order by 
+                table_name
+        '''
         
-        columns = pd.read_csv(self.download("select table_name, column_name, udt_name from information_schema.columns where table_name != 'notificacao' order by table_name", join(self.tables,'columns.csv')))
+        columns_sql = '''
+            select 
+                table_name, column_name, udt_name 
+            from 
+                information_schema.columns 
+            where 
+                table_name != 'notificacao' 
+            order by 
+                table_name
+        '''
         
-        constraint = pd.read_csv(self.download("SELECT tc.constraint_name, tc.table_name, kcu.column_name, ccu.table_name AS foreign_table_name, ccu.column_name AS foreign_column_name FROM information_schema.table_constraints AS tc JOIN information_schema.key_column_usage AS kcu ON tc.constraint_name = kcu.constraint_name JOIN information_schema.constraint_column_usage AS ccu ON ccu.constraint_name = tc.constraint_name WHERE constraint_type = 'FOREIGN KEY'",join(self.tables,'notifica_constraint.csv')))
+        constraint_sql = '''
+            SELECT
+                tc.constraint_name, tc.table_name, kcu.column_name,
+                ccu.table_name AS foreign_table_name,
+                ccu.column_name AS foreign_column_name
+            FROM
+                information_schema.table_constraints AS tc
+                JOIN information_schema.key_column_usage AS kcu
+                    ON tc.constraint_name = kcu.constraint_name
+                JOIN information_schema.constraint_column_usage AS ccu
+                    ON ccu.constraint_name = tc.constraint_name
+            WHERE constraint_type = 'FOREIGN KEY'
+        '''
+        
+        notificacao = pd.read_csv(self.download(notificacao_sql, join(self.tables,'notificacao.csv')))
+        
+        columns = pd.read_csv(self.download(columns_sql, join(self.tables,'columns.csv')))
+        
+        constraint = pd.read_csv(self.download(constraint_sql,join(self.tables,'notifica_constraint.csv')))
     
         notificacao = pd.merge(notificacao,constraint,on=['table_name','column_name'],how='left').fillna('')
         
@@ -61,17 +111,17 @@ class Metabase:
             raise Exception(f"Query {query_name}.sql not found in {self.sql_dir}")
 
         with open(join(self.sql_dir,f"{query_name}.sql")) as file:
-            raw_sql = trim_overspace(" ".join([ row.replace('\n',' ') for row in file.readlines() if not '--' in row ])).lower()
+            raw_sql = trim_overspace(" ".join([ row.replace('\n',' ') for row in file.readlines() if not '--' in row ]))
 
         sql = {}
-        sql['select'] = re.search('select (.*) from', raw_sql).group(1)
-        sql['from'] = re.search('from (.*) where', raw_sql).group(1)
-        sql['where'] = re.search('where (.*) order', raw_sql).group(1)
-        sql['order'] = re.search('order (.*) limit', raw_sql).group(1)
-        sql['limit'] = re.search('limit (.*) offset', raw_sql).group(1)
-        sql['offset'] = re.search('offset (.*)', raw_sql).group(1)
+        sql['select'] = re.search('SELECT (.*) FROM', raw_sql).group(1)
+        sql['from'] = re.search('FROM (.*) WHERE', raw_sql).group(1)
+        sql['where'] = re.search('WHERE (.*) ORDER', raw_sql).group(1)
+        sql['order'] = re.search('ORDER (.*) LIMIT', raw_sql).group(1)
+        sql['limit'] = re.search('LIMIT (.*) OFFSET', raw_sql).group(1)
+        sql['offset'] = re.search('OFFSET (.*)', raw_sql).group(1)
         
-        query_size = pd.read_csv(self.download(f"select count(*) from {sql['from']} where {sql['where']}", join(self.output,f"{query_name}_len.csv"))).iloc[0,0]
+        query_size = pd.read_csv(self.download(f"select count(*) from {sql['from']} where {sql['where']}", join(self.output,f"{query_name}_len.csv"),False)).iloc[0,0]
         print(f"\nquery_size: {query_size}\n")
                 
         parts = []
@@ -84,25 +134,23 @@ class Metabase:
                 )
             )
 
-#         print(f"\n{parts}\n")
         output_path = join(self.output,f"{query_name}.csv")
     
-        output_query = pd.concat([ pd.read_csv(part, low_memory=False) for part in parts ])
-        print(f"saving in {output_path}")
-        output_query.to_csv(output_path,index=False)
-        
-        if remove_parts:
-            print(f"removing {' '.join(parts)}")
-            for part in parts: os.remove(part) 
+        if len(parts) > 1: 
+            output_query = pd.concat([ pd.read_csv(part) for part in parts ])
+            print(f"saving all in {output_path}")
+            output_query.to_csv(output_path,index=False)
+            if remove_parts:
+                print(f"removing {' '.join(parts)}")
+                for part in parts: os.remove(part)         
+        else:
+            os.rename(parts[0],output_path)
         
         return output_path     
-                                                                                             
-    def __request_cookie(self):
-        return input('enter with cookie')                                                                  
-                                                                                             
+                                
                                                                                              
     @retry(Exception, delay=10, tries=-1)
-    def download(self,sql,pathfile):
+    def download(self,sql,pathfile, log=True):
         
         header = {
             'Host':'metabase.saude.pr.gov.br',
@@ -114,7 +162,7 @@ class Metabase:
             'Origin':'https://metabase.saude.pr.gov.br',
             'Connection':'keep-alive',
             'Referer':'https://metabase.saude.pr.gov.br/question',
-            'Cookie': self.cookie,
+            'Cookie': f"metabase.SESSION={self.cookie}",
             'Upgrade-Insecure-Requests':'1'
         }
         
@@ -137,24 +185,23 @@ class Metabase:
                             )
 
         if res.status_code in [ 200, 202 ] :
-            print(f"Success code {res.status_code }")
+            if log: print(f"Success code {res.status_code }")
         else:
-            print(f"Error code {res.status_code }")
+            if log: print(f"Error code {res.status_code }")
             raise Exception()
 
         with open(pathfile,'wb') as out:
-            print(f'Saving in {pathfile}')
+#             if log: print(f'Saving in {pathfile}')
             for chunk in res.iter_content(chunk_size=1024):
                 if chunk:
                     out.write(chunk)
                     out.flush()
 
         try:
-            print(f"Download finish, time elapsed: {res.elapsed}")
-            print(f"downloaded shape {pd.read_csv(pathfile,low_memory=False).shape}\n")
+            if log: print(f"Download finish, time elapsed: {res.elapsed}")
+            if log: print(f"downloaded shape {pd.read_csv(pathfile,low_memory=False).shape}\n")
         except:
-            print(f"download error")
-            raise Exception()
+            raise Exception(f"download error")
 
 
         return pathfile
